requirements.txt
PyPDF2
python-docx
transformers
torch
faiss-cpu  # 如果使用 CPU 版本；如果使用 GPU 版本，可替换为 faiss-gpu
openai
numpy
toml